requirements.txt
requests
Google-api-python-client
python-dotenv
groq
langchain
langchain-core
langchain-groq
rich
gTTS
elevenlabs
moviepy
numpy
nltk